+++ package.json
@@ -41,6 +41,7 @@
     "ci:gha": "turbo run ci:gha",
     "ci": "turbo run ci",
     "ci:publish": "turbo run ci:publish",
+    "version": "bun run tools/version.ts",
     "clean": "git clean -fdX -e '.env' -e '.env.*' -e '!.env' -e '!.env.*' && find . -type f \\( -name 'pnpm-lock.yaml' -o -name 'package-lock.json' -o -name 'yarn.lock' -o -name 'bun.lockb' -o -name 'bun.lock' \\) -exec rm -f {} + && bun install",
     "dev": "turbo dev",
     "lint": "turbo lint",
